refactor: reduce circular dependencies
