+++ development/circular-deps.jsonc
@@ -56,6 +56,5 @@
     "ui/store/actions.ts"
   ],
   ["ui/ducks/metamask/metamask.js", "ui/selectors/selectors.js"],
-  ["ui/ducks/metamask/metamask.js", "ui/store/actions.ts"],
-  ["ui/selectors/multichain.ts", "ui/selectors/selectors.js"]
+  ["ui/ducks/metamask/metamask.js", "ui/store/actions.ts"]
 ]
