+++ ui/helpers/utils/multichain/blockExplorer.test.ts
@@ -10,7 +10,7 @@ import {
   NETWORK_TYPES,
   CURRENCY_SYMBOLS,
 } from '../../../../shared/constants/network';
-import { MultichainNetwork } from '../../../selectors/multichain';
+import { MultichainNetwork } from '../../../selectors/multichain/networks';
 import {
   getMultichainBlockExplorerUrl,
   getMultichainAccountUrl,
