+++ ui/helpers/utils/multichain/blockExplorer.ts
@@ -1,6 +1,6 @@
 import { KnownCaipNamespace, parseCaipChainId } from '@metamask/utils';
 import { getAccountLink } from '@metamask/etherscan-link';
-import { MultichainNetwork } from '../../../selectors/multichain';
+import { MultichainNetwork } from '../../../selectors/multichain/networks';
 // TODO: Remove restricted import
 // eslint-disable-next-line import/no-restricted-paths
 import { normalizeSafeAddress } from '../../../../app/scripts/lib/multichain/address';
